Cleaned up main.c
Moved some functions to other files to cleanup main.c
