+++ Arkanoid/include/Ball.h
@@ -1,6 +1,11 @@
 #ifndef _Ball_H_
 #define _Ball_H_
 
+struct BallPos{
+    long x,y;
+	long xV, yV;
+};
+
 void initBall();
 void drawBall(struct BallPos * ball);
